Refactor validation.rs
Split this file up into separate modules to make it easier to navigate and
add new validator types.
