+++ crates/spk-solve/crates/validation/src/lib.rs
@@ -8,13 +8,8 @@
 mod error;
 mod impossible_checks;
 mod validation;
+pub mod validators;
 
 pub use error::{Error, Result};
 pub use impossible_checks::{ImpossibleRequestsChecker, IMPOSSIBLE_CHECKS_TARGET};
-pub use validation::{
-    default_validators,
-    BinaryOnlyValidator,
-    GetMergedRequest,
-    ValidatorT,
-    Validators,
-};
+pub use validation::{default_validators, GetMergedRequest, ValidatorT, Validators};
